+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## 3.0.2 - 2020-01-14
+### Fixed
+- DB Dump now appropriately removes backup files from storage/backups folder
+- Changed failing functions from php die command to Craft Error logging
+
 ## 3.0.1 - 2019-11-08
 ### Fixed
 - Allowed anonymous access to controller actions